requirements.txt
# Core
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
tqdm>=4.60.0
Pillow>=8.3.0

# Streamlit app
streamlit>=1.0.0

# Optional / utilities
opencv-python>=4.5.0
scipy>=1.7.0

# For model exporting / saving
onnx>=1.12.0

# Notes: GPU-specific installs (CUDA) should match your PyTorch build.
